requirements.txt
APScheduler==3.10.1
blinker==1.6.2
click==8.1.6
decorator==5.1.1
Flask==2.3.2
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzlocal==5.0.1
validators==0.20.0
Werkzeug==2.3.6
